app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.21
loguru==0.5.3
requests==2.31.0
sqlite3==3.41.2
aiohttp==3.8.5
python-telegram-bot

